Create requireAuth higher order component for authentication
