+++ client/src/actions/index.js
@@ -9,6 +9,7 @@ export const signup = (email, password, callback) => async dispatch => {
             'password': password
         });
         dispatch({ type: AUTH_USER, payload: response.data.token });
+        localStorage.setItem('token', response.data.token);
         callback();
     } catch(e) {   
         dispatch({ type: AUTH_ERROR, payload: "Email in use"});
