+++ client/src/components/Profile.js
@@ -1,25 +1,15 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { withRouter } from 'react-router-dom';
 import * as actions from '../actions';
+import requireAuth from './requireAuth';
 
 class Profile extends Component {
-    state = { name: '' };
-
-    // Our component just got rendered
-    componentDidMount() {
-        this.shouldNavigateAway();
-    }
-
-    // Our component just got updated
-    componentDidUpdate() {
-        this.shouldNavigateAway();
-    }
-
-    shouldNavigateAway() {
-        if(!this.props.auth) {
-            console.log("I need to leave!");
-            this.props.history.push('/');
+    constructor(props) {
+        super(props);
+        this.state = {
+            name: "",
+            bio: "",
+            photo: "",
         }
     }
 
@@ -27,15 +17,10 @@ class Profile extends Component {
         return(
             <div style={{display: 'flex', flexDirection: 'column'}}>
                 <span>This is the profile component</span>
-                <span>{this.props.uuid}</span>
-                <span>{this.props.token}</span>
+                <span>{this.props.auth}</span>
             </div>
         );
     }
 }
 
-function mapStateToProps(state) {
-    return { auth: state.auth.authenticated };
-}
-
-export default withRouter(connect(mapStateToProps, actions)(Profile));
+export default connect(null, actions)(requireAuth(Profile));
