+++ client/src/index.js
@@ -10,7 +10,9 @@ import reducers from './reducers';
 
 const store = createStore(
   reducers,
-  {},
+  { 
+    auth: { authenticated: localStorage.getItem('token') }
+  },
   applyMiddleware(reduxThunk)
 );
